# contracts/http/public.openapi.yaml
openapi: 3.0.3
info:
  title: IELTS Learning System Public API
  version: 1.0.0
  description: >-
    Canonical learner/client HTTP machine contract for the Go Core public boundary.
    The contract transports canonical product semantics but does not redefine Band,
    evidence sufficiency, Progression, planner ranking, content quality, or product
    coverage policy. Durable learner/product truth remains in Core.
security:
  - ClerkBearer: []
paths:
...
  /v1/me:
    get:
      operationId: getMe
      summary: Read the authenticated Core actor projection
      responses:
        '200':
          description: Learner-safe identity projection for the authenticated actor.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalFailure'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /v1/target-profile:
    get:
      operationId: getTargetProfile
      summary: Read the current TargetProfile
      description: >-
        Missing target input is represented as a normal unresolved/not-configured domain
        result. It is not learner evidence insufficiency and is not a CoverageGap.
      responses:
        '200':
          description: Current TargetProfile state, including an ordinary not-configured result.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetProfileReadResult'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalFailure'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
    put:
      operationId: putTargetProfile
      summary: Replace the current TargetProfile
      description: >-
        PUT is a full replacement of supplied TargetProfile constraints. Omitted optional
        request fields mean that constraint is not supplied in the resulting profile.
        The resource-revision precondition is decisive with the mutation. A value of 0
        means the caller expects no TargetProfile to exist; a positive value must equal
        the current resource revision. Partial targets are valid domain state, while
        target-relative planning/readiness/support remains unresolved until canonical
        resolution requirements are met.
      parameters:
        - $ref: '#/components/parameters/ExpectedResourceRevision'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutTargetProfileRequest'
      responses:
        '200':
          description: Existing TargetProfile replaced.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetProfile'
        '201':
          description: TargetProfile created while the expected revision was 0.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetProfile'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '412':
          $ref: '#/components/responses/StaleResourceRevision'
        '422':
          $ref: '#/components/responses/SemanticPreconditionFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalFailure'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /v1/daily-plan:
    get:
      operationId: getDailyPlan
      summary: Read the current DailyPlan snapshot
      description: >-
        The DailyPlan is a snapshot and never an assignment grant. Starting a referenced
        activity requires current hard-eligibility and assignment checks. Unresolved
        target conditions and product CoverageGaps remain separate from learner gaps.
      responses:
        '200':
          description: Current plan snapshot, including valid unresolved/coverage-blocked states.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyPlan'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalFailure'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /v1/practice-modes:
    get:
      operationId: listPracticeModes
      summary: List the stable learner-facing practice-mode catalog
      description: >-
        Returns catalog identity/packaging only. Catalog presence does not assert that a
        mode is currently assignable for this learner, target, product-support state, or
        evidence consequence.
      responses:
        '200':
          description: Small stable canonical practice-mode catalog projection.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeModeList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalFailure'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /v1/practice-activities:
    post:
      operationId: createPracticeActivity
      summary: Resolve and assign one current eligible PracticeActivity
      description: >-
        Current hard eligibility is resolved at assignment time. A DailyPlan reference
        cannot force stale assignment. Accepted/replayed assignment is keyed by the same
        Idempotency-Key and actor/operation scope; the same key with an incompatible
        request fails closed. Valid learner-specific unavailability is a domain result,
        not an infrastructure failure. Learner-facing activity material never includes
        hidden answer/model/rubric/evaluator material before reveal is permitted.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePracticeActivityRequest'
      responses:
        '200':
          description: >-
            Idempotent replay of the authoritative logical result, or a valid domain
            result explaining why no current assignment is available.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeActivityCreationResult'
        '201':
          description: A new exact eligible ContentRevision was authoritatively assigned.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeActivityCreationResult'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/SemanticPreconditionFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalFailure'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
        '504':
          $ref: '#/components/responses/AmbiguousOutcome'
  /v1/practice-activities/{practice_activity_id}:
    parameters:
      - $ref: '#/components/parameters/PracticeActivityId'
    get:
      operationId: getPracticeActivity
      summary: Read one learner-owned PracticeActivity projection
      responses:
        '200':
          description: >-
            Exact assigned revision and learner-safe material. Hidden answer/model/rubric
            material remains absent unless the owning reveal policy permits it.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeActivity'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFoundOrNotVisible'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalFailure'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /v1/practice-activities/{practice_activity_id}/media/{media_reference}:
    get:
      operationId: getPracticeActivityMedia
      summary: Read one exact learner-owned PracticeActivity media reference
      description: >-
        Returns the immutable media bytes referenced by an assigned learner-owned activity.
        The reference is opaque and is never interpreted as a filesystem path.
      parameters:
        - $ref: '#/components/parameters/PracticeActivityId'
        - $ref: '#/components/parameters/MediaReference'
      responses:
        '200':
          description: Exact immutable Ogg bytes for the assigned activity revision.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            audio/ogg:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFoundOrNotVisible'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalFailure'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /v1/attempts:
    post:
      operationId: createAttempt
      summary: Create a draft Attempt against an assigned PracticeActivity
      description: >-
        The Attempt binds the exact assigned ContentRevision. Creation is idempotent
        because duplicate retry would duplicate learner history.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttemptRequest'
      responses:
        '200':
          description: Idempotent replay of the same authoritative draft Attempt.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
        '201':
          description: New draft Attempt created.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFoundOrNotVisible'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/SemanticPreconditionFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalFailure'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /v1/attempts/{attempt_id}:
    parameters:
      - $ref: '#/components/parameters/AttemptId'
    get:
      operationId: getAttempt
      summary: Read one learner-owned Attempt
      responses:
        '200':
          description: Current authoritative Attempt state.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFoundOrNotVisible'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalFailure'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
    patch:
      operationId: patchAttempt
      summary: Replace mutable draft response state on an Attempt
      description: >-
        Only a draft Attempt may be mutated. The expected resource revision is compared
        atomically with the guarded mutation. Submitted/evaluating/evaluated/invalid
        Attempts are immutable for this Attempt identity; retry/redraft creates another
        related Attempt under owning product semantics rather than rewriting history.
      parameters:
        - $ref: '#/components/parameters/ExpectedResourceRevision'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchAttemptRequest'
      responses:
        '200':
          description: Draft response updated and resource revision advanced.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFoundOrNotVisible'
        '409':
          $ref: '#/components/responses/StateConflict'
        '412':
          $ref: '#/components/responses/StaleResourceRevision'
        '422':
          $ref: '#/components/responses/SemanticPreconditionFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalFailure'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
...
components:
  securitySchemes:
    ClerkBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Short-lived Clerk-issued token supplied as `Authorization: Bearer <token>`.
        Go Core verifies the selected Clerk integration security conditions and maps the
        authenticated external principal to a stable Core actor. Clerk roles, permissions,
        organizations, and metadata are not product authorization authority; Core owns
        capabilities, entitlements, and all product authorization.
  headers:
    RequestId:
      description: Privacy-safe request correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    InvalidRequest:
      x-failure-class: OPERATION_REJECTED
      x-error-code: INVALID_REQUEST
      description: Request framing, syntax, or machine-contract validation failed.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthenticated:
      x-failure-class: OPERATION_REJECTED
      x-error-code: UNAUTHENTICATED
      description: A valid authenticated public principal could not be established.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      x-failure-class: OPERATION_REJECTED
      x-error-code: FORBIDDEN
      description: Authenticated actor lacks permission for the requested capability/scope.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RateLimited:
      x-failure-class: OPERATION_REJECTED
      x-error-code: RATE_LIMITED
      x-retry-requirement: RETRY_AFTER
      description: Transport/abuse/capacity admission rejected this request temporarily.
      headers:
        Retry-After:
          description: Retry delay where the server can safely provide one.
          schema:
            type: integer
            minimum: 0
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    DependencyUnavailable:
      x-failure-class: INFRASTRUCTURE_FAILURE
      x-error-code: DEPENDENCY_UNAVAILABLE
      description: >-
        Required infrastructure cannot currently establish the intended authoritative
        result. This is never encoded as learner weakness, CoverageGap, or score zero.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalFailure:
      x-failure-class: INFRASTRUCTURE_FAILURE
      x-error-code: INTERNAL_FAILURE
      description: >-
        Core encountered an internal infrastructure/runtime failure and did not convert it
        into learner evidence, weakness, CoverageGap, or score.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    Me:
      type: object
      additionalProperties: false
      required: [actor_id, learner_id]
      properties:
        actor_id:
          $ref: '#/components/schemas/ResourceId'
        learner_id:
          $ref: '#/components/schemas/ResourceId'
...